refactor(Backend): Model-Klassen sind nun nach Entwurfsmuster in entsprechenden Packeten organisiert
