+++ backend/src/main/java/de/fhdortmund/kidsapp/model/Fabrik/RegistrierterNutzer.java
@@ -1,8 +1,11 @@
-package de.fhdortmund.kidsapp.model;
+package de.fhdortmund.kidsapp.model.Fabrik;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import de.fhdortmund.kidsapp.model.AnschriftT;
+import de.fhdortmund.kidsapp.model.Nutzer;
+import de.fhdortmund.kidsapp.model.Veranstaltung;
 import jakarta.persistence.*;
 import lombok.Data;
