+++ backend/src/main/java/de/fhdortmund/kidsapp/model/Kompositum/MultiChoice.java
@@ -1,4 +1,4 @@
-package de.fhdortmund.kidsapp.model;
+package de.fhdortmund.kidsapp.model.Kompositum;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,11 +9,13 @@ import jakarta.persistence.FetchType;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-
+/**
+ * Entity representing a multiple-choice question type.
+ */
 @Entity
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class SingleChoice extends Fragentyp {
+public class MultiChoice extends Fragentyp {
     @ElementCollection(fetch = FetchType.EAGER)
     private List<String> antwortoptionen = new ArrayList<>();
 }
