+++ backend/src/main/java/de/fhdortmund/kidsapp/model/Zustaende/Stattfindend.java
@@ -1,7 +1,7 @@
-package de.fhdortmund.kidsapp.model;
+package de.fhdortmund.kidsapp.model.Zustaende;
 
+import de.fhdortmund.kidsapp.model.Veranstaltung;
 import jakarta.persistence.Entity;
-import jakarta.persistence.OneToOne;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
