+++ backend/src/main/java/de/fhdortmund/kidsapp/repository/UmfrageRepository.java
@@ -3,7 +3,7 @@ package de.fhdortmund.kidsapp.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import de.fhdortmund.kidsapp.model.Umfrage;
+import de.fhdortmund.kidsapp.model.Kompositum.Umfrage;
 
 
 @Repository
